summarizer/requirements.txt
fastapi
pydantic
uvicorn
langchain
langchain-core
langchain_openai
pytube
python-dotenv
openai
pytest